+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Added: REPLACE() string function (see issue #560)
 * Added: NEWID(), UUID() and GEN_RANDOM_UUID() functions for GUID
 * Added: DEFAULT for column can be a function (e.g. DEFAULT NEWID())
+* Changed: SELECT * FROM ? GROUP BY a works as FIRST(*)
 
 ### 0.2.2 "Mitilini" (15.01.2016)
 * Fix: SELECT can use functions from GROUP BY list